visualization2/server/requirements.txt
certifi==2020.6.20
click==6.7
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
ImageHash==4.2.0
imageio==2.9.0
numpy==1.19.0
opencv-contrib-python==4.2.0.34
Pillow==7.2.0
scikit-image==0.17.2
scipy==1.5.4
